+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.9.0] - 2026-04-05
+
 ### Added
 
 - **exports**: `Plugin`, `InfiniteDepthConfigWithExtends`, and